Remove the head node and tail node
If the head node is NULL or the queue is empty, it will return.
By using list_first_entry and list_last_entry from list.h,
it is easier to get the whole object memory space to free.
These two functions rely on the important macro container_of,
which calculates the address of the structure containing the address
pointer.

Change-Id: Iee39326926994fd84e57a19d482bf2a9bf11ce00