Deletion of middle of the queue
If the queue is NULL or empty, it will return. I use two pointers
to find the middle node: one pointer moves forward from the head of
the queue, while the other moves backward. Because it's a double-
linked list, both pointers must meet somewhere. When we find the
node to delete, we use the macro list_entry to free the entire
address space of the node.

Change-Id: I2e20c17cd24854f681a7c818cb3a4bc076d8b375